refactor: Move checkbox tests to collocated checkboxes.test.ts
