+++ server/editor/index.test.ts
@@ -89,7 +89,7 @@ b. Do that.`;
 
   // Find the ordered_list in the result
   const orderedList = json?.content?.find(
-    (node: any) => node.type === "ordered_list"
+    (node: { type: string }) => node.type === "ordered_list"
   );
 
   expect(orderedList).toBeDefined();
@@ -144,64 +144,4 @@ test("serializes uppercase alpha lists back to markdown", () => {
   const output = serializer.serialize(ast);
 
   expect(output.trim()).toBe("A. First item\nB. Second item");
-});
-
-test("preserves mixed checkbox and regular items in a list", () => {
-  const markdown = `- [x] Checked item
-- Regular item
-- [ ] Unchecked item`;
-
-  const ast = parser.parse(markdown);
-  const json = ast?.toJSON();
-
-  const checkboxList = json?.content?.find(
-    (node: any) => node.type === "checkbox_list"
-  );
-
-  expect(checkboxList).toBeDefined();
-  expect(checkboxList?.content).toHaveLength(3);
-  expect(checkboxList?.content[0].type).toBe("checkbox_item");
-  expect(checkboxList?.content[1].type).toBe("checkbox_item");
-  expect(checkboxList?.content[2].type).toBe("checkbox_item");
-});
-
-test("round-trips mixed checkbox lists through serializer", () => {
-  const markdown = `- [x] Checked
-- Plain text
-- [ ] Unchecked`;
-
-  const ast = parser.parse(markdown);
-  const output = serializer.serialize(ast);
-
-  // All items should survive the round-trip
-  expect(output).toContain("Checked");
-  expect(output).toContain("Plain text");
-  expect(output).toContain("Unchecked");
-});
-
-test("does not convert nested bullet list items inside checkbox lists", () => {
-  const markdown = `- [x] Parent checkbox
-    - Nested bullet item
-    - Another nested item
-- [ ] Second checkbox`;
-
-  const ast = parser.parse(markdown);
-  const json = ast?.toJSON();
-
-  const checkboxList = json?.content?.find(
-    (node: any) => node.type === "checkbox_list"
-  );
-
-  expect(checkboxList).toBeDefined();
-  expect(checkboxList?.content).toHaveLength(2);
-  expect(checkboxList?.content[0].type).toBe("checkbox_item");
-  expect(checkboxList?.content[1].type).toBe("checkbox_item");
-
-  // Nested list should remain a bullet_list, not a checkbox_list
-  const nestedContent = checkboxList?.content[0].content;
-  const nestedList = nestedContent?.find(
-    (node: any) => node.type === "bullet_list"
-  );
-  expect(nestedList).toBeDefined();
-  expect(nestedList?.content[0].type).toBe("list_item");
 });
